+++ src/app/board/_components/BulletinContainer.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import Bulletin from "./Bulletin";
+import BestBulletin from "./BestBulletin";
+
+function BulletinContainer() {
+  return (
+    <div className="m-4 flex flex-col gap-6 md:m-6 w-full max-w-300">
+      <BestBulletin />
+      <Bulletin />
+    </div>
+  );
+}
+
+export default BulletinContainer;
